[refact] lib/common.py 함수 정리 및 메일 관련 코드 분리
- 사용하지 않는 함수 삭제
- 메일 상수 & mailer 함수 분리 (lib/mail.py)
