+++ admin/admin_sendmail.py
@@ -4,9 +4,10 @@ from fastapi import APIRouter, Depends, Request, Form
 
 from core.template import AdminTemplates
 from lib.common import (
-    domain_mail_host, get_admin_email, get_admin_email_name, mailer
+    domain_mail_host, get_admin_email, get_admin_email_name
 )
 from lib.dependency.dependencies import validate_token
+from lib.mail import mailer
 
 router = APIRouter()
 templates = AdminTemplates()
