+++ bbs/formmail.py
@@ -7,7 +7,9 @@ from core.database import db_session
 from core.exception import AlertException
 from core.models import Member
 from core.template import UserTemplates
-from lib.common import captcha_widget, mailer, StringEncrypt
+from lib.captcha import captcha_widget
+from lib.common import StringEncrypt
+from lib.mail import mailer
 from lib.dependency.dependencies import validate_token
 
 router = APIRouter()
